requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
pyyaml>=6.0

# Data storage
pyarrow>=10.0.0  # For parquet support

# Optional but recommended
python-dateutil>=2.8.0
pytz>=2023.3
